Add ToOkHsv and ToOkHsl methods for Color in C#
It builds, but I can't seem to get Godot working with C# on my end, so I can't really test whether it works.
It's not very efficient, converting the color 3 times for one call, but I'm not sure this minor performance improvement would be worth the hassle, so I'm committing it as is.
Especially since GDScript doesn't do it efficiently either.